New activity starts
Added second activity and have it working with sub topics been sent to
the new activity.
This is a change over the last commit. There was a change from the full
topic to the string array of subtopics been sent as it was easier to
implement as it caused less issues with the transfer of information.
It should be noted that if a topic does not find subtopics the topics
then we cause a null pointer exception which will cause the application
to crash.