Fixed move methods so that they now move right way
the down and up were moving the square around the wrong way but i fixed
it. Also I added in the other square to see if there was any problems
with renderer both at the same time. It crashed the first time,
panicked, commented out some lines and reintroduced them and now there
is no crashing for whatever reason.